Extract reusable `FeedStoreSpecs` helper methods into a shared scope to be used by other `FeedStore` implementations Tests
