+++ EssentialFeedTests/Feed Cache/CodableFeedStoreTests.swift
@@ -180,59 +180,6 @@ final class CodableFeedStoreTests: XCTestCase, FailableFeedStore {
         return sut
     }
     
-    @discardableResult
-    private func insert(_ cache: (feed: [LocalFeedImage], timestamp: Date), to sut: FeedStore) -> Error? {
-        let exp = expectation(description: "Wait for cache retrieval")
-        
-        var insertionError: Error?
-        sut.insert(cache.feed, timestamp: cache.timestamp) { receivedInsertionError in
-            insertionError = receivedInsertionError
-            exp.fulfill()
-        }
-        
-        wait(for: [exp], timeout: 1.0)
-        return insertionError
-    }
-    
-    private func deleteCache (from sut: FeedStore) -> Error? {
-        let exp = expectation (description: "Wait for cache deletion")
-        
-        var deletionError: Error?
-        sut.deleteCachedFeed { receivedDeletionError in
-            deletionError = receivedDeletionError
-            exp.fulfill()
-        }
-        
-        wait (for: [exp], timeout: 3.0)
-        return deletionError
-    }
-    
-    private func expect(_ sut: FeedStore, toRetrieve expectedResult: RetrieveCachedFeedResult, file: StaticString = #file, line: UInt = #line) {
-        let exp = expectation(description: "Wait for cache retrieval")
-        
-        sut.retrieve { result in
-            switch (expectedResult, result) {
-                case (.empty, .empty),
-                    (.failure, .failure):
-                    break
-                case let (.found(expected), .found(retrieved)):
-                    XCTAssertEqual(retrieved.feed, expected.feed, file: file, line: line)
-                    XCTAssertEqual(retrieved.timestamp, expected.timestamp, file: file, line: line)
-                default:
-                    XCTFail("Expected to retrieve \(expectedResult), got \(result) instead", file: file, line: line)
-            }
-            
-            exp.fulfill()
-        }
-        
-        wait(for: [exp], timeout: 1.0)
-    }
-    
-    private func expect(_ sut: FeedStore, toRetrieveTwice expectedResult: RetrieveCachedFeedResult, file: StaticString = #file, line: UInt = #line) {
-        expect(sut, toRetrieve: expectedResult, file: file, line: line)
-        expect(sut, toRetrieve: expectedResult, file: file, line: line)
-    }
-    
     private func setupEmptyStoreState() {
         deleteStoreArtifacts()
     }
